Ajustei a estrutura do projeto.
