+++ Projeto_Concessionária/nbproject/project.properties
@@ -71,7 +71,7 @@ jlink.additionalmodules=
 jlink.additionalparam=
 jlink.launcher=true
 jlink.launcher.name=Projeto_Concession_ria
-main.class=projeto_concession\u00e1ria.Projeto_Concession\u00e1ria
+main.class=projeto_concessionaria.ProjetoConcessionaria
 manifest.file=manifest.mf
 meta.inf.dir=${src.dir}/META-INF
 mkdist.disabled=false
