--- Projeto_Concessionária/src/painelconcessionaria/views/PainelConcessionaria.java
@@ -1,20 +1,0 @@
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Main.java to edit this template
- */
-package painelconcessionaria.views;
-
-/**
- *
- * @author USER
- */
-public class PainelConcessionaria {
-
-    /**
-     * @param args the command line arguments
-     */
-    public static void main(String[] args) {
-        // TODO code application logic here
-    }
-    
-}
