+++ Projeto_Concessionária/src/projeto_concessionaria/controller/FuncionarioController.java
@@ -2,11 +2,11 @@
  * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
  * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
  */
-package projeto_concessionário.Controller;
+package projeto_concessionaria.controller;
 
 import java.util.ArrayList;
 import java.util.List;
-import projeto_concessionária.Model.Funcionario;
+import projeto_concessionaria.Model.Funcionario;
 
 
 /**
