+++ Projeto_Concessionária/src/projeto_concessionaria/model/Carro.java
@@ -2,21 +2,23 @@
  * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
  * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
  */
-package projeto_concessionária.Model;
+package projeto_concessionaria.Model;
 
-import projeto_concessionária.Model.enums.Status;
-import projeto_concessionária.Model.enums.TipoCombustivel;
+import projeto_concessionaria.model.enums.Status;
+import projeto_concessionaria.model.enums.TipoCombustivel;
 
 /**
  *
  * @author Guilherme
  */
-public class Moto extends Veiculo {
+public class Carro extends Veiculo{
+    
     private Integer quantidadeDeRodas;
 
-    public Moto(String marca, String modelo, String placa, Integer anoFabricacao, String cor, TipoCombustivel tipoCombustivel, Double quilometragem, Status status, String chassi, double preco) {
+    public Carro(String marca, String modelo, String placa, Integer anoFabricacao, String cor, TipoCombustivel tipoCombustivel, Double quilometragem, Status status, String chassi,Integer quantidadeDeRodas,double preco) {
         super(marca, modelo, placa, anoFabricacao, cor, tipoCombustivel, quilometragem, status, chassi,preco);
         this.quantidadeDeRodas = quantidadeDeRodas;
+        
     }
 
     public Integer getQuantidadeDeRodas() {
@@ -26,12 +28,6 @@ public class Moto extends Veiculo {
     public void setQuantidadeDeRodas(Integer quantidadeDeRodas) {
         this.quantidadeDeRodas = quantidadeDeRodas;
     }
-
-    
-
-
-    
-
