+++ Projeto_Concessionária/src/projeto_concessionaria/model/Funcionario.java
@@ -2,9 +2,10 @@
  * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
  * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
  */
-package projeto_concessionária.Model;
+package projeto_concessionaria.Model;
 
-import projeto_concessionária.Model.enums.CargoFuncionario;
+import projeto_concessionaria.model.enums.CargoFuncionario;
+
 
 /**
  *
