+++ Projeto_Concessionária/src/projeto_concessionaria/model/Moto.java
@@ -2,23 +2,21 @@
  * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
  * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
  */
-package projeto_concessionária.Model;
+package projeto_concessionaria.Model;
 
-import projeto_concessionária.Model.enums.Status;
-import projeto_concessionária.Model.enums.TipoCombustivel;
+import projeto_concessionaria.model.enums.Status;
+import projeto_concessionaria.model.enums.TipoCombustivel;
 
 /**
  *
  * @author Guilherme
  */
-public class Carro extends Veiculo{
-    
+public class Moto extends Veiculo {
     private Integer quantidadeDeRodas;
 
-    public Carro(String marca, String modelo, String placa, Integer anoFabricacao, String cor, TipoCombustivel tipoCombustivel, Double quilometragem, Status status, String chassi,Integer quantidadeDeRodas,double preco) {
+    public Moto(String marca, String modelo, String placa, Integer anoFabricacao, String cor, TipoCombustivel tipoCombustivel, Double quilometragem, Status status, String chassi, double preco) {
         super(marca, modelo, placa, anoFabricacao, cor, tipoCombustivel, quilometragem, status, chassi,preco);
         this.quantidadeDeRodas = quantidadeDeRodas;
-        
     }
 
     public Integer getQuantidadeDeRodas() {
@@ -28,6 +26,12 @@ public class Carro extends Veiculo{
     public void setQuantidadeDeRodas(Integer quantidadeDeRodas) {
         this.quantidadeDeRodas = quantidadeDeRodas;
     }
+
+    
+
+
+    
+
