+++ Projeto_Concessionária/src/projeto_concessionaria/model/MovimentacaoEstoque.java
@@ -2,7 +2,7 @@
  * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
  * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
  */
-package projeto_concessionária.Model;
+package projeto_concessionaria.Model;
 
 import java.util.Date;
