+++ Projeto_Concessionária/src/projeto_concessionaria/model/enums/CargoFuncionario.java
@@ -2,7 +2,7 @@
  * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
  * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Enum.java to edit this template
  */
-package projeto_concessionária.Model.enums;
+package projeto_concessionaria.model.enums;
 
 /**
  *
