+++ Projeto_Concessionária/src/projeto_concessionaria/model/enums/Status.java
@@ -2,17 +2,13 @@
  * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
  * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Enum.java to edit this template
  */
-package projeto_concessionária.Model.enums;
+package projeto_concessionaria.model.enums;
 
 /**
  *
  * @author Guilherme
  */
-public enum TipoCombustivel {
-    GASOLINA,
-    ETANOL,
-    DIESEL,
-    FLEX,
-    ELETRICO,
-    HIBRIDO;
-    }
+public enum Status {
+    DISPONIVEL,
+    MANUTENCAO;
+}
