+++ Projeto_Concessionária/src/projeto_concessionaria/model/enums/TipoCombustivel.java
@@ -2,13 +2,17 @@
  * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
  * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Enum.java to edit this template
  */
-package projeto_concessionária.Model.enums;
+package projeto_concessionaria.model.enums;
 
 /**
  *
  * @author Guilherme
  */
-public enum Status {
-    DISPONIVEL,
-    MANUTENCAO;
-}
+public enum TipoCombustivel {
+    GASOLINA,
+    ETANOL,
+    DIESEL,
+    FLEX,
+    ELETRICO,
+    HIBRIDO;
+    }
