+++ Projeto_Concessionária/src/projeto_concessionaria/view/Painel.java
@@ -1,3 +1,5 @@
+package projeto_concessionaria.view;
+
 /*
  * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
  * Click nbfs://nbhost/SystemFileSystem/Templates/GUIForms/JFrame.java to edit this template
