+++ Projeto_Concessionária/src/projeto_concessionaria/view/PainelConcessionaria.java
@@ -1,0 +1,20 @@
+/*
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Main.java to edit this template
+ */
+package projeto_concessionaria.view;
+
+/**
+ *
+ * @author USER
+ */
+public class PainelConcessionaria {
+
+    /**
+     * @param args the command line arguments
+     */
+    public static void main(String[] args) {
+        // TODO code application logic here
+    }
+    
+}
