+++ Projeto_Concessionária/src/projeto_concessionaria/view/TelaMovimentacoes.java
@@ -2,7 +2,7 @@
  * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
  * Click nbfs://nbhost/SystemFileSystem/Templates/GUIForms/JPanel.java to edit this template
  */
-package painelconcessionaria.views;
+package projeto_concessionaria.view;
 
 /**
  *
